conversion accuracy

+++ tsib/src/task_readApps.h
@@ -9,7 +9,8 @@
 #include "atomtimer.h"
 
 uint8_t apps_ch;
-uint16_t tempApp;
+uint32_t tempApp;
+
 void apps_adcinit(void);
 void task_readApps(uint32_t data);
